Update phpDoc comments to include all standard attributes.
The addition of @license, @author, @copyright and @version in all
files that do not already have these attributes or have them
incorrectly listed. Also, the addition of several other comments
to functions in core libraries, though there are still some of
these that need to be done at a later date (such as decorators).